Refactoring: cxx tests
