+++ openjij/main.cpp
@@ -1,4 +1,4 @@
-//    Copyright 2021 Jij Inc.
+//    Copyright 2023 Jij Inc.
 
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
