+++ tests/cxxtest.cpp
@@ -24,14 +24,16 @@
 #include <openjij/utility/gpu/cublas.hpp>
 #include <openjij/sampler/sa_sampler.hpp>
 
-
 #include <Eigen/Dense>
 #include <Eigen/Sparse>
 
+#define TEST_CASE_INDEX 1
+#include "testcase.hpp"
+
 #include "polynomial_test.hpp"
-#include "./cxxtest/binary_polynomial_model.hpp"
-#include "./cxxtest/binary_polynomial_sa_system.hpp"
-#include "./cxxtest/binary_polynomial_sa_sampler.hpp"
+#include "./cxxtest/sampler/all.hpp"
+#include "./cxxtest/system/all.hpp"
+#include "./cxxtest/graph/all.hpp"
 
 // #####################################
 // helper functions
@@ -42,411 +44,13 @@
  * @return classical interaction which represents specific optimization problem
  */
 
-static constexpr std::size_t num_system_size = 8;
 
-#define TEST_CASE_INDEX 1
-
-#include "testcase.hpp"
-
-
-static openjij::utility::ClassicalScheduleList generate_schedule_list(){
-    return openjij::utility::make_classical_schedule_list(0.1, 100.0, 100, 100);
-}
-
-static openjij::utility::TransverseFieldScheduleList generate_tfm_schedule_list(){
-    return openjij::utility::make_transverse_field_schedule_list(10, 100, 100);
-}
 
 // #####################################
 // tests
 // #####################################
 
 #include <chrono>
-
-//speed test
-//TEST(Graph, speedtest){
-//    using namespace openjij::graph;
-//    std::size_t N = 5000;
-//    auto begin = std::chrono::high_resolution_clock::now();
-//    Dense<double> a(N);
-//    for(std::size_t i=0; i<N; i++){
-//        for(std::size_t j=i; j<N; j++){
-//            a.J(i, j)  = 1;
-//        }
-//    }
-//    auto end = std::chrono::high_resolution_clock::now();
-//    std::cout << "time: " << std::chrono::duration_cast<std::chrono::milliseconds>(end-begin).count() << std::endl;
-//
-//    auto quad = cimod::Quadratic<size_t, double>();
-//    auto lin = cimod::Linear<size_t, double>();
-//    begin = std::chrono::high_resolution_clock::now();
-//    for(std::size_t i=0; i<N; i++){
-//        for(std::size_t j=i+1; j<N; j++){
-//            quad[std::make_pair(i,j)] = 1;
-//        }
-//    }
-//    for(std::size_t i=0; i<N; i++){
-//        lin[i] = 1;
-//    }
-//    auto bqm = cimod::BinaryQuadraticModel<size_t, double>(lin, quad, 0, cimod::Vartype::BINARY);
-//    end = std::chrono::high_resolution_clock::now();
-//    std::cout << "time: " << std::chrono::duration_cast<std::chrono::milliseconds>(end-begin).count() << std::endl;
-//}
-
-TEST(Graph, DenseGraphCheck){
-    using namespace openjij::graph;
-    using namespace openjij;
-
-    std::size_t N = 500;
-    Dense<double> a(N);
-    auto r = utility::Xorshift(1234);
-    auto urd = std::uniform_real_distribution<>{-10, 10};
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i; j<N; j++){
-            a.J(i, j)  = urd(r);
-        }
-    }
-
-    r = utility::Xorshift(1234);
-
-    // check if graph holds correct variables
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i; j<N; j++){
-            EXPECT_EQ(a.J(i, j) , urd(r));
-        }
-    }
-
-    r = utility::Xorshift(1234);
-
-    // check if graph index is reversible (Jij = Jji)
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i; j<N; j++){
-            EXPECT_EQ(a.J(j, i) , urd(r));
-        }
-    }
-}
-
-TEST(Graph, SparseGraphCheck){
-    using namespace openjij::graph;
-    using namespace openjij;
-
-    std::size_t N = 500;
-    Sparse<double> b(N, N-1);
-    auto r = utility::Xorshift(1234);
-    auto urd = std::uniform_real_distribution<>{-10, 10};
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i+1; j<N; j++){
-            b.J(i, j) = urd(r);
-        }
-    }
-
-    r = utility::Xorshift(1234);
-
-    // check if graph holds correct variables
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i+1; j<N; j++){
-            EXPECT_EQ(b.J(i, j) , urd(r));
-        }
-    }
-
-    r = utility::Xorshift(1234);
-
-    // check if graph index is reversible (Jij = Jji)
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i+1; j<N; j++){
-            EXPECT_EQ(b.J(j, i) , urd(r));
-        }
-    }
-
-    //check adj_nodes
-    for(std::size_t i=0; i<N; i++){
-        std::size_t tot = 0;
-        for(auto&& elem : b.adj_nodes(i)){
-            tot += elem;
-        }
-        EXPECT_EQ(tot, N*(N-1)/2 - i);
-    }
-    EXPECT_EQ(b.get_num_edges(), N-1);
-
-    Sparse<double> c(N, N);
-    r = utility::Xorshift(1234);
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i; j<N; j++){
-            c.J(j, i) = urd(r);
-        }
-    }
-
-    r = utility::Xorshift(1234);
-
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i; j<N; j++){
-            EXPECT_EQ(c.J(i, j) , urd(r));
-        }
-    }
-
-    r = utility::Xorshift(1234);
-
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i; j<N; j++){
-            EXPECT_EQ(c.J(j, i) , urd(r));
-        }
-    }
-    for(std::size_t i=0; i<N; i++){
-        std::size_t tot = 0;
-        for(auto&& elem : c.adj_nodes(i)){
-            tot += elem;
-        }
-        EXPECT_EQ(tot, N*(N-1)/2);
-    }
-    EXPECT_EQ(c.get_num_edges(), N);
-}
-
-TEST(Graph, EnergyCheck){
-    using namespace openjij::graph;
-    std::size_t N = 500;
-
-    Dense<double> b_d(N);
-    Sparse<double> b(N, N-1);
-
-    Spins spins(N, 1);
-    Spins spins_neg(N, -1);
-    auto random_engine = std::mt19937(1);
-    Spins spins_r = b_d.gen_spin(random_engine);
-
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i+1; j<N; j++){
-            b_d.J(i, j) = 1;
-        }
-    }
-
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i+1; j<N; j++){
-            b.J(i, j) = 1;
-        }
-    }
-
-    // generate interaction matrix from dense graph
-    Eigen::SparseMatrix<double, Eigen::RowMajor> mat = b_d.get_interactions().sparseView();
-    // and generate CSRSparse
-    CSRSparse<double> b_csrs(mat.template triangularView<Eigen::Upper>());
-
-    EXPECT_EQ(b_d.calc_energy(spins), (1./2) * (N*N - N));
-    EXPECT_EQ(b_d.calc_energy(spins_neg), (1./2) * (N*N - N));
-    EXPECT_EQ(b.calc_energy(spins), (1./2) * (N*N - N));
-    EXPECT_EQ(b.calc_energy(spins_neg), (1./2) * (N*N - N));
-    EXPECT_EQ(b_csrs.calc_energy(spins), (1./2) * (N*N - N));
-    EXPECT_EQ(b_csrs.calc_energy(spins_neg), (1./2) * (N*N - N));
-    EXPECT_EQ(b_d.calc_energy(spins_r), b.calc_energy(spins_r));
-    EXPECT_EQ(b_d.calc_energy(spins_r), b_csrs.calc_energy(spins_r));
-
-    Dense<double> c_d(N);
-    Sparse<double> c(N, N);
-
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i; j<N; j++){
-            c_d.J(i, j) = 1;
-        }
-    }
-
-    for(std::size_t i=0; i<N; i++){
-        for(std::size_t j=i; j<N; j++){
-            c.J(i, j) = 1;
-        }
-    }
-    // generate interaction matrix from dense graph
-    mat = c_d.get_interactions().sparseView();
-    // and generate CSRSparse
-    CSRSparse<double> c_csrs(mat.template triangularView<Eigen::Upper>());
-
-    EXPECT_EQ(c_d.calc_energy(spins), (1./2) * (N*N + N));
-    EXPECT_EQ(c_d.calc_energy(spins_neg), (1./2) * (N*N - 3*N));
-    EXPECT_EQ(c.calc_energy(spins), (1./2) * (N*N + N));
-    EXPECT_EQ(c.calc_energy(spins_neg), (1./2) * (N*N - 3*N));
-    EXPECT_EQ(c_csrs.calc_energy(spins), (1./2) * (N*N + N));
-    EXPECT_EQ(c_csrs.calc_energy(spins_neg), (1./2) * (N*N - 3*N));
-    EXPECT_EQ(c_d.calc_energy(spins_r), c.calc_energy(spins_r));
-    EXPECT_EQ(c_d.calc_energy(spins_r), c_csrs.calc_energy(spins_r));
-}
-
-//json tests
-TEST(Graph, JSONTest){
-    using namespace cimod;
-    using namespace openjij;
-
-    Linear<uint32_t, double> linear{ {0, 1.0}, {11, 3.0}};
-    Quadratic<uint32_t, double> quadratic
-    {
-        {std::make_pair(0, 4), 12.0}, {std::make_pair(4, 12), 13.0}, {std::make_pair(6, 14), 14.0},
-            {std::make_pair(3, 4), 23.0}, {std::make_pair(11, 12), 24.0},
-            {std::make_pair(5, 13), 34.0}
-    };
-    double offset = 0.0;
-    Vartype vartype = Vartype::SPIN;
-    BinaryQuadraticModel<uint32_t, double, cimod::Sparse> bqm_k4(linear, quadratic, offset, vartype);
-    auto s = graph::Chimera<double>(bqm_k4.to_serializable(), 1, 2);
-    EXPECT_NEAR(s.J(0,1,3,graph::ChimeraDir::IN_0or4), 24, 1e-5);
-    EXPECT_NEAR(s.J(0,0,4,graph::ChimeraDir::PLUS_C), 13, 1e-5);
-    EXPECT_NEAR(s.J(0,1,6,graph::ChimeraDir::MINUS_C), 14, 1e-5);
-    EXPECT_NEAR(s.J(0,0,4,graph::ChimeraDir::IN_3or7), 23, 1e-5);
-    EXPECT_NEAR(s.J(0,0,3,graph::ChimeraDir::IN_0or4), 23, 1e-5);
-    EXPECT_NEAR(s.J(0,1,5,graph::ChimeraDir::MINUS_C), 34, 1e-5);
-    EXPECT_NEAR(s.h(0,1,3), 3, 1e-5);
-}
-
-//ClassicalIsing tests
-
-TEST(ClassicalIsing, GenerateTheSameEigenObject){
-    using namespace openjij;
-    graph::Dense<double> d(4);
-    graph::Sparse<double> s(4);
-    d.J(2,3) = s.J(2,3) = 4;
-    d.J(1,0) = s.J(1,0) = -2;
-    d.J(1,1) = s.J(1,1) = 5;
-    d.J(2,2) = s.J(2,2) = 10;
-
-    auto engine_for_spin = std::mt19937(1);
-    auto cl_dense = system::make_classical_ising(d.gen_spin(engine_for_spin), d);
-    auto cl_sparse = system::make_classical_ising(s.gen_spin(engine_for_spin), s);
-    Eigen::MatrixXd m1 = cl_dense.interaction;
-    //convert from sparse to dense
-    Eigen::MatrixXd m2 = cl_sparse.interaction;
-    EXPECT_EQ(m1, m2);
-}
-
-//TODO: macro?
-//SingleSpinFlip tests
-
-
-TEST(SingleSpinFlip, FindTrueGroundState_ClassicalIsing_Dense) {
-    using namespace openjij;
-
-    //generate classical dense system
-    const auto interaction = generate_interaction<graph::Dense<double>>();
-    auto engine_for_spin = std::mt19937(1);
-    const auto spin = interaction.gen_spin(engine_for_spin);
-    auto classical_ising = system::make_classical_ising(spin, interaction); 
-    
-    auto random_numder_engine = std::mt19937(1);
-    const auto schedule_list = generate_schedule_list();
-
-    algorithm::Algorithm<updater::SingleSpinFlip>::run(classical_ising, random_numder_engine, schedule_list);
-
-    EXPECT_EQ(get_true_groundstate(), result::get_solution(classical_ising));
-}
-
-TEST(SingleSpinFlip, FindTrueGroundState_ClassicalIsing_Sparse) {
-    using namespace openjij;
-
-    //generate classical dense system
-    const auto interaction = generate_interaction<graph::Sparse<double>>();
-    auto engine_for_spin = std::mt19937(1);
-    const auto spin = interaction.gen_spin(engine_for_spin);
-    auto classical_ising = system::make_classical_ising(spin, interaction);
-    
-    auto random_numder_engine = std::mt19937(1);
-    const auto schedule_list = generate_schedule_list();
-
-    algorithm::Algorithm<updater::SingleSpinFlip>::run(classical_ising, random_numder_engine, schedule_list);
-
-    EXPECT_EQ(get_true_groundstate(), result::get_solution(classical_ising));
-}
-
-TEST(SingleSpinFlip, FindTrueGroundState_ClassicalIsing_CSRSparse) {
-    using namespace openjij;
-
-    //generate classical dense system
-    const auto dense_interaction = generate_interaction<graph::Dense<double>>();
-    //output sparse interaction
-    Eigen::SparseMatrix<double, Eigen::RowMajor> sp_mat = dense_interaction.get_interactions().sparseView();
-    const auto interaction = graph::CSRSparse<double>(sp_mat.template triangularView<Eigen::Upper>());
-    auto engine_for_spin = std::mt19937(1);
-    const auto spin = interaction.gen_spin(engine_for_spin);
-    auto classical_ising = system::make_classical_ising(spin, interaction);
-    
-    auto random_numder_engine = std::mt19937(1);
-    const auto schedule_list = generate_schedule_list();
-
-    algorithm::Algorithm<updater::SingleSpinFlip>::run(classical_ising, random_numder_engine, schedule_list);
-
-    EXPECT_EQ(get_true_groundstate(), result::get_solution(classical_ising));
-}
-
-TEST(SingleSpinFlip, FindTrueGroundState_TransverseIsing_Dense) {
-    using namespace openjij;
-
-    //generate classical dense system
-    const auto interaction = generate_interaction<graph::Dense<double>>();
-    auto engine_for_spin = std::mt19937(1);
-    std::size_t num_trotter_slices = 10;
-
-    //generate random trotter spins
-    system::TrotterSpins init_trotter_spins(num_trotter_slices);
-    for(auto& spins : init_trotter_spins){
-        spins = interaction.gen_spin(engine_for_spin);
-    }
-
-    auto transverse_ising = system::make_transverse_ising(init_trotter_spins, interaction, 1.0);
-
-    auto transverse_ising2 = system::make_transverse_ising(interaction.gen_spin(engine_for_spin), interaction, 1.0, 10);
-    
-    auto random_numder_engine = std::mt19937(1);
-    const auto schedule_list = generate_tfm_schedule_list();
-
-    algorithm::Algorithm<updater::SingleSpinFlip>::run(transverse_ising, random_numder_engine, schedule_list);
-
-    EXPECT_EQ(get_true_groundstate(), result::get_solution(transverse_ising));
-}
-
-TEST(SingleSpinFlip, FindTrueGroundState_TransverseIsing_Sparse) {
-    using namespace openjij;
-
-    //generate classical dense system
-    const auto interaction = generate_interaction<graph::Sparse<double>>();
-    auto engine_for_spin = std::mt19937(1);
-    std::size_t num_trotter_slices = 10;
-
-    //generate random trotter spins
-    system::TrotterSpins init_trotter_spins(num_trotter_slices);
-    for(auto& spins : init_trotter_spins){
-        spins = interaction.gen_spin(engine_for_spin);
-    }
-
-    auto transverse_ising = system::make_transverse_ising(init_trotter_spins, interaction, 1.0); //gamma = 1.0
-    
-    auto random_numder_engine = std::mt19937(1);
-    const auto schedule_list = generate_tfm_schedule_list();
-
-    algorithm::Algorithm<updater::SingleSpinFlip>::run(transverse_ising, random_numder_engine, schedule_list);
-
-    EXPECT_EQ(get_true_groundstate(), result::get_solution(transverse_ising));
-}
-
-TEST(SingleSpinFlip, FindTrueGroundState_TransverseIsing_CSRSparse) {
-    using namespace openjij;
-
-    //generate classical dense system
-    const auto dense_interaction = generate_interaction<graph::Dense<double>>();
-    //output sparse interaction
-    Eigen::SparseMatrix<double, Eigen::RowMajor> sp_mat = dense_interaction.get_interactions().sparseView();
-    const auto interaction = graph::CSRSparse<double>(sp_mat.template triangularView<Eigen::Upper>());
-    auto engine_for_spin = std::mt19937(1);
-    std::size_t num_trotter_slices = 10;
-
-    //generate random trotter spins
-    system::TrotterSpins init_trotter_spins(num_trotter_slices);
-    for(auto& spins : init_trotter_spins){
-        spins = interaction.gen_spin(engine_for_spin);
-    }
-
-    auto transverse_ising = system::make_transverse_ising(init_trotter_spins, interaction, 1.0); //gamma = 1.0
-    
-    auto random_numder_engine = std::mt19937(1);
-    const auto schedule_list = generate_tfm_schedule_list();
-
-    algorithm::Algorithm<updater::SingleSpinFlip>::run(transverse_ising, random_numder_engine, schedule_list);
-
-    EXPECT_EQ(get_true_groundstate(), result::get_solution(transverse_ising));
-}
