+++ tests/cxxtest/graph/binary_polynomial_model.hpp
