+++ tests/cxxtest/system/classical_ising.hpp
@@ -1,0 +1,39 @@
+//    Copyright 2023 Jij Inc.
+
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+
+//        http://www.apache.org/licenses/LICENSE-2.0
+
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+
+#pragma once
+
+namespace openjij {
+namespace test {
+
+TEST(ClassicalIsing, GenerateTheSameEigenObject){
+    using namespace openjij;
+    graph::Dense<double> d(4);
+    graph::Sparse<double> s(4);
+    d.J(2,3) = s.J(2,3) = 4;
+    d.J(1,0) = s.J(1,0) = -2;
+    d.J(1,1) = s.J(1,1) = 5;
+    d.J(2,2) = s.J(2,2) = 10;
+
+    auto engine_for_spin = std::mt19937(1);
+    auto cl_dense = system::make_classical_ising(d.gen_spin(engine_for_spin), d);
+    auto cl_sparse = system::make_classical_ising(s.gen_spin(engine_for_spin), s);
+    Eigen::MatrixXd m1 = cl_dense.interaction;
+    //convert from sparse to dense
+    Eigen::MatrixXd m2 = cl_sparse.interaction;
+    EXPECT_EQ(m1, m2);
+}
+
+}
+}
